requirements.txt

flask==2.3.3
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.2.1
